streamline module and fix broken tests
