+++ MANIFEST.in
@@ -1,1 +1,2 @@
 include requirements.txt
+recursive-include taskcat/project_templates *
