+++ taskcat/__init__.py
@@ -9,3 +9,4 @@ from taskcat.client_factory import ClientFactory
 from taskcat.exceptions import TaskCatException
 from taskcat.s3_sync import S3Sync
 from taskcat.lambda_build import LambdaBuild
+from taskcat.project_generator import ProjectConfiguration, ProjectGenerator
