+++ taskcat/project_templates/quickstart/README.md.jinja
