+++ taskcat/project_templates/quickstart/ci/debug-input.json.jinja
