+++ taskcat/project_templates/quickstart/ci/taskcat-autobucket-json.yml.jinja
