+++ taskcat/project_templates/quickstart/ci/taskcat-autobucket.yml.jinja
