+++ taskcat/project_templates/quickstart/ci/taskcat.yml.jinja
