+++ taskcat/project_templates/quickstart/scripts/scripts_userdata.sh.jinja
