+++ taskcat/project_templates/quickstart/templates/debug-yaml.template.jinja
