+++ taskcat/project_templates/quickstart/templates/debug.template.jinja
